move main tests to cmd
